Added enums for future functionality - easier to add now. Added a 404 exception to display to front end if id doesn't exist

+++ backend/src/main/java/com/kurzic/reading_app_backend/services/BookService.java
@@ -3,6 +3,7 @@ package com.kurzic.reading_app_backend.services;
 import com.kurzic.reading_app_backend.DTOs.BookRequestDTO;
 import com.kurzic.reading_app_backend.DTOs.BookResponseDTO;
 import com.kurzic.reading_app_backend.entities.Book;
+import com.kurzic.reading_app_backend.exceptions.BookNotFoundException;
 import com.kurzic.reading_app_backend.repositories.BookRepository;
 import org.springframework.stereotype.Service;
 
@@ -23,6 +24,7 @@ public class BookService {
         book.setAuthor(newBook.author());
         book.setIsbn(newBook.isbn());
         book.setPageCount(newBook.pageCount());
+        book.setStatus(newBook.status());
 
         Book savedBook = repo.save(book);
         return mapToResponse(savedBook);
@@ -36,26 +38,30 @@ public class BookService {
     }
 
     public BookResponseDTO getBookById(Long id) {
-        Book book = repo.findById(id).orElse(null);
-        if (book == null) return null;
+        Book book = repo.findById(id)
+                .orElseThrow(() -> new BookNotFoundException(id));
 
         return mapToResponse(book);
     }
 
     public BookResponseDTO updateBook(Long id, BookRequestDTO updatedBook) {
-        Book existing = repo.findById(id).orElse(null);
-        if (existing == null) return null;
+        Book book = repo.findById(id)
+                .orElseThrow(() -> new  BookNotFoundException(id));
 
-        existing.setTitle(updatedBook.title());
-        existing.setAuthor(updatedBook.author());
-        existing.setPageCount(updatedBook.pageCount());
-        existing.setIsbn(updatedBook.isbn());
+        book.setTitle(updatedBook.title());
+        book.setAuthor(updatedBook.author());
+        book.setPageCount(updatedBook.pageCount());
+        book.setIsbn(updatedBook.isbn());
+        book.setStatus(updatedBook.status());
 
-        Book savedBook = repo.save(existing);
+        Book savedBook = repo.save(book);
         return mapToResponse(savedBook);
     }
 
-    public void deleteBookByID(Long id){
+    public void deleteBookByID(Long id) {
+        if (!repo.existsById(id)) {
+            throw new BookNotFoundException(id);
+        }
         repo.deleteById(id);
     }
 
@@ -65,7 +71,8 @@ public class BookService {
                 book.getTitle(),
                 book.getAuthor(),
                 book.getIsbn(),
-                book.getPageCount()
+                book.getPageCount(),
+                book.getStatus()
         );
     }
 }
